requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-genai>=1.0.0
fal-client>=0.5.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.97.0
elevenlabs>=1.0.0
yt-dlp>=2025.1.0
python-docx>=1.0.0
